Unidad 5

+++ UNIDAD 5/Ejercicio3(Alumno-Maestro)/Persona.cs
@@ -11,7 +11,7 @@ namespace Ejercicio3_Alumno_Maestro_
         public string nombre { get; set; }
         public DateTime fechaNacimiento { get; set; }
         public string curp { get; set; }
-        public int telefono { get; set; }
+        public long telefono { get; set; }
         public string eMail { get; set; }
     }
 }
